# book-network-ui/src/app/services/models/change-password-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChangePasswordRequest {
  confirmPassword?: string;
  currentPassword?: string;
  newPassword?: string;
}