# book-network-ui/src/app/services/models/notification-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationResponse {
  bookTitle?: string;
  createdAt?: string;
  id?: number;
  message?: string;
  read?: boolean;
  status?: 'BORROWED' | 'RETURNED' | 'RETURN_APPROVED' | 'RESERVED' | 'CANCELLED';
}