# book-network-ui/src/app/services/models/user-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserRequest {
  dateOfBirth: string;
  email: string;
  firstName: string;
  lastName: string;
}